requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
ecdsa==0.19.1
fastapi==0.128.0
h11==0.16.0
idna==3.11
passlib==1.7.4
pillow==12.1.0
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-jose==3.5.0
python-multipart==0.0.22
reportlab==4.4.9
requests==2.32.5
rsa==4.9.1
six==1.17.0
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
websockets==16.0

# PostgreSQL Database Dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
